AIHackathon/requirements_new.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
aiohttp==3.9.1
PyPDF2==3.0.1
python-docx==1.0.1
pandas==2.1.3
python-magic==0.4.27
python-dateutil==2.8.2
Jinja2==3.1.2
pathlib==1.0.1
PyYAML==6.0.1